requirements.txt
accelerate
anytree
bitsandbytes
datasets
ipywidgets
jupyterlab
langchain
langchain-anthropic
langchain-chroma
langchain-core
langchain-google-genai
langchain-huggingface
langchain-ollama
langchain-openai
langgraph
matplotlib
mlx
mlx_lm
numpy
pandas
peft
python-dotenv
scikit-learn
seaborn
tqdm
transformers
